requirements.txt
Flask==2.3.2
flask-cors==3.0.10
PyPDF2==3.0.1
Pillow==10.0.0
python-docx==0.8.11

To install the dependencies listed, run:
bash
pip install -r requirements.txt